api/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
Requests==2.32.2
uvicorn